# swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: API Absensi - Admin & Super Admin Endpoints
  description: |
    Dokumentasi API untuk endpoints Admin dan Super Admin sistem absensi.
    
    **Level Akses:**
    - Super Admin (Level 1): Akses penuh ke semua endpoint
    - Admin OPD (Level 2): Akses terbatas ke SKPD yang ditugaskan
    - Admin UPT (Level 3): Akses terbatas ke Satker yang ditugaskan
    
    **Autentikasi:** Semua endpoint memerlukan JWT token dengan level admin yang sesuai.
  version: 1.0.0
  contact:
    name: API Support
    email: support@absensi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/admin
    description: Development Server
  - url: https://api.absensi.com/api/admin
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token untuk autentikasi admin
...
paths:
...
  /skpd/{kdskpd}/satker:
    get:
      tags:
        - Organization Hierarchy
      summary: Get Satker by SKPD
      description: Mendapatkan Satker berdasarkan SKPD
      security:
        - bearerAuth: []
      parameters:
        - name: kdskpd
          in: path
          required: true
          description: Kode SKPD
          schema:
            type: string
            example: "01"
      responses:
        '200':
          description: List of Satker under SKPD
  /skpd/{kdskpd}/satker/{kdsatker}/bidang:
    get:
      tags:
        - Organization Hierarchy
      summary: Get Bidang by Satker
      description: Mendapatkan Bidang berdasarkan Satker
      security:
        - bearerAuth: []
      parameters:
        - name: kdskpd
          in: path
          required: true
          description: Kode SKPD
          schema:
            type: string
            example: "01"
        - name: kdsatker
          in: path
          required: true
          description: Kode Satker
          schema:
            type: string
            example: "001"
      responses:
        '200':
          description: List of Bidang under Satker

  # Jadwal Kegiatan Endpoints
...